Set up dependency injection for services in the front end and included a temporary TestService to show setup + functionality
