--- frontend/component/copyright.tsx
@@ -1,17 +1,0 @@
-import {Link, Typography} from "@mui/material";
-import * as React from "react";
-
-function Copyright(props: any) {
-    return (
-        <Typography variant="body2" color="text.secondary" align="center" {...props}>
-            {'Copyright © '}
-            <Link color="inherit" href="https://techatncsu.github.io/">
-                Tech@NCSU
-            </Link>{' '}
-            {new Date().getFullYear()}
-            {'.'}
-        </Typography>
-    );
-}
-
-export default Copyright;
