+++ frontend/pages/index.tsx
@@ -1,12 +1,15 @@
 import type { NextPage } from 'next'
-import {useTestService} from "../services/TestService";
+import {useTestService} from "../src/services/TestService";
 
 const Home: NextPage = () => {
 
   const testService = useTestService();
 
   return(
-      <div>{testService.getTestPhrase()}</div>
+      <>
+          <div>Price of Bitcoin: {testService.price}</div>
+          <button onClick={testService.testMethod}>Show Test Method</button>
+      </>
   );
 }
