+++ frontend/src/services/ProvidedServices.ts
@@ -1,3 +1,6 @@
+/**
+ * List of services used by the application
+ */
 enum ProvidedServices { TestService }
 
 export default ProvidedServices;
